Remove truncate_messages calls from various nodes and add CustomChatOpenAI class for improved message handling
