+++ prometheus/app/services/llm_service.py
@@ -3,26 +3,21 @@ from typing import Optional
 from langchain_anthropic import ChatAnthropic
 from langchain_core.language_models.chat_models import BaseChatModel
 from langchain_google_genai import ChatGoogleGenerativeAI
-from langchain_openai import ChatOpenAI
 
-
-class CustomChatOpenAI(ChatOpenAI):
-    def bind_tools(self, tools, tool_choice=None, **kwargs):
-        kwargs["parallel_tool_calls"] = False
-        return super().bind_tools(tools, tool_choice=tool_choice, **kwargs)
+from prometheus.chat_models.custom_chat_openai import CustomChatOpenAI
 
 
 class LLMService:
     def __init__(
-        self,
-        advanced_model_name: str,
-        base_model_name: str,
-        openai_format_api_key: Optional[str] = None,
-        openai_format_base_url: Optional[str] = None,
-        anthropic_api_key: Optional[str] = None,
-        gemini_api_key: Optional[str] = None,
-        temperature: float = 0.0,
-        max_output_tokens: int = 15000,
+            self,
+            advanced_model_name: str,
+            base_model_name: str,
+            openai_format_api_key: Optional[str] = None,
+            openai_format_base_url: Optional[str] = None,
+            anthropic_api_key: Optional[str] = None,
+            gemini_api_key: Optional[str] = None,
+            temperature: float = 0.0,
+            max_output_tokens: int = 15000,
     ):
         self.advanced_model = get_model(
             advanced_model_name,
@@ -45,13 +40,13 @@ class LLMService:
 
 
 def get_model(
-    model_name: str,
-    openai_format_api_key: Optional[str] = None,
-    openai_format_base_url: Optional[str] = None,
-    anthropic_api_key: Optional[str] = None,
-    gemini_api_key: Optional[str] = None,
-    temperature: float = 0.0,
-    max_output_tokens: int = 15000,
+        model_name: str,
+        openai_format_api_key: Optional[str] = None,
+        openai_format_base_url: Optional[str] = None,
+        anthropic_api_key: Optional[str] = None,
+        gemini_api_key: Optional[str] = None,
+        temperature: float = 0.0,
+        max_output_tokens: int = 15000,
 ) -> BaseChatModel:
     if "claude" in model_name:
         return ChatAnthropic(
@@ -70,6 +65,9 @@ def get_model(
             max_retries=3,
         )
     else:
+        """
+        Use tiktoken_counter to ensure that the input messages do not exceed the maximum token limit.
+        """
         return CustomChatOpenAI(
             model=model_name,
             api_key=openai_format_api_key,
