+++ prometheus/lang_graph/nodes/context_provider_node.py
@@ -16,7 +16,6 @@ from langchain_core.messages import SystemMessage
 
 from prometheus.graph.knowledge_graph import KnowledgeGraph
 from prometheus.tools import graph_traversal
-from prometheus.utils.lang_graph_util import truncate_messages
 
 
 class ContextProviderNode:
@@ -351,8 +350,7 @@ Available AST node types for code structure search: {ast_node_types}
         """
         self._logger.debug(f"Context provider messages: {state['context_provider_messages']}")
         message_history = [self.system_prompt] + state["context_provider_messages"]
-        truncated_message_history = truncate_messages(message_history)
-        response = self.model_with_tools.invoke(truncated_message_history)
+        response = self.model_with_tools.invoke(message_history)
         self._logger.debug(response)
         # The response will be added to the bottom of the list
         return {"context_provider_messages": [response]}
