+++ prometheus/utils/llm_util.py
@@ -1,0 +1,37 @@
+from typing import Sequence
+
+import tiktoken
+from langchain_core.messages import BaseMessage, HumanMessage, AIMessage, ToolMessage, SystemMessage
+from langchain_core.output_parsers import StrOutputParser
+
+
+def str_token_counter(text: str) -> int:
+    enc = tiktoken.get_encoding("o200k_base")
+    return len(enc.encode(text))
+
+
+def tiktoken_counter(messages: Sequence[BaseMessage]) -> int:
+    """Approximately reproduce https://github.com/openai/openai-cookbook/blob/main/examples/How_to_count_tokens_with_tiktoken.ipynb
+
+    For simplicity only supports str Message.contents.
+    """
+    output_parser = StrOutputParser()
+    num_tokens = 3  # every reply is primed with <|start|>assistant<|message|>
+    tokens_per_message = 3
+    tokens_per_name = 1
+    for msg in messages:
+        if isinstance(msg, HumanMessage):
+            role = "user"
+        elif isinstance(msg, AIMessage):
+            role = "assistant"
+        elif isinstance(msg, ToolMessage):
+            role = "tool"
+        elif isinstance(msg, SystemMessage):
+            role = "system"
+        else:
+            raise ValueError(f"Unsupported messages type {msg.__class__}")
+        msg_content = output_parser.invoke(msg)
+        num_tokens += tokens_per_message + str_token_counter(role) + str_token_counter(msg_content)
+        if msg.name:
+            num_tokens += tokens_per_name + str_token_counter(msg.name)
+    return num_tokens
